requirements.txt
amqp==5.0.6
asgiref==3.4.1
astroid==2.9.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==4.2.4
celery==5.2.0b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.10
django-celery-beat==2.2.1
django-timezone-field==4.2.1
google-api-core==2.3.1
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gspread==5.0.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
isort==5.10.1
kombu==5.0.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.1
pandas==1.1.5
platformdirs==2.4.0
prompt-toolkit==3.0.23
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyline==0.3.20
pylint==2.12.1
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rope==0.22.0
rsa==4.8
six==1.16.0
sodapy==2.1.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
